# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API quản lý nhà hàng, nhân viên, thực đơn, đơn hàng, file upload...",
        "title": "Restaurant Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/shifts": {
            "get": {
                "description": "Lấy danh sách ca làm việc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Lấy danh sách ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trang",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Số lượng mỗi trang",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Shift"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Tạo mới một ca làm việc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Tạo mới ca làm việc",
                "parameters": [
                    {
                        "description": "Dữ liệu ca làm việc",
                        "name": "shift",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Shift"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Shift"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shifts-chedules": {
            "get": {
                "description": "Lấy danh sách lịch ca làm việc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shiftschedule"
                ],
                "summary": "Lấy danh sách lịch ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trang",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Số lượng mỗi trang",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ShiftSchedule"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Tạo mới một hoặc nhiều lịch ca làm việc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shiftschedule"
                ],
                "summary": "Tạo mới lịch ca làm việc",
                "parameters": [
                    {
                        "description": "Dữ liệu lịch ca làm việc",
                        "name": "shiftSchedules",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ShiftSchedule"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ShiftSchedule"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shifts-chedules/{id}": {
            "get": {
                "description": "Lấy chi tiết một lịch ca làm việc theo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shiftschedule"
                ],
                "summary": "Lấy chi tiết lịch ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID shift schedule",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ShiftSchedule"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa một lịch ca làm việc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shiftschedule"
                ],
                "summary": "Xóa lịch ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID shift schedule",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Cập nhật thông tin lịch ca làm việc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shiftschedule"
                ],
                "summary": "Cập nhật lịch ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID shift schedule",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dữ liệu cập nhật",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shifts/{id}": {
            "get": {
                "description": "Lấy chi tiết một ca làm việc theo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Lấy chi tiết ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID shift",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Shift"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa một ca làm việc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Xóa ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID shift",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Cập nhật thông tin ca làm việc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Cập nhật ca làm việc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID shift",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dữ liệu cập nhật",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enum.EmployeeScheduleType": {
            "type": "string",
            "enum": [
                "FullTime",
                "PartTime"
            ],
            "x-enum-varnames": [
                "FullTime",
                "PartTime"
            ]
        },
...
        "model.Account": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                },
                "role_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "model.Employee": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/model.Account"
                },
                "account_id": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "avatar_file": {
                    "$ref": "#/definitions/model.File"
                },
                "avatar_file_id": {
                    "type": "integer"
                },
                "base_salary": {
                    "type": "integer"
                },
                "birthday": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "join_date": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "salary_per_hour": {
                    "type": "integer"
                },
                "schedule_type": {
                    "$ref": "#/definitions/enum.EmployeeScheduleType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.File": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mime_type": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Permission": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "permission_name": {
                    "type": "string"
                }
            }
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Permission"
                    }
                },
                "role_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Shift": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "shift_name": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.ShiftSchedule": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "employee": {
                    "$ref": "#/definitions/model.Employee"
                },
                "employee_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "shift": {
                    "$ref": "#/definitions/model.Shift"
                },
                "shift_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        }
    }
}